Feat: added async endpoints using fastapi on backend

+++ app/backend/crud/employee/crud.py
@@ -1,49 +1,47 @@
-from sqlalchemy.orm import Session
+from sqlalchemy.ext.asyncio import AsyncSession
+from sqlalchemy.future import select
 from models.employee.employee_schema import EmployeeUpdate, EmployeeCreate
 from models.employee.employee import EmployeeModel
 
-
-def get_employee(db: Session, employee_id: int):
+async def get_employee(db: AsyncSession, employee_id: int):
     """
     Função que recebe um id e retorna somente o funcionário correspondente
     """
-    return db.query(EmployeeModel).filter(EmployeeModel.employee_id == employee_id).first()
+    result = await db.execute(select(EmployeeModel).where(EmployeeModel.employee_id == employee_id))
+    return result.scalars().first()
 
-
-def get_employees(db: Session):
+async def get_employees(db: AsyncSession):
     """
     Função que retorna todos os funcionários
     """
-    return db.query(EmployeeModel).all()
+    result = await db.execute(select(EmployeeModel))
+    return result.scalars().all()
 
-
-def create_employee(db: Session, employee: EmployeeCreate):
+async def create_employee(db: AsyncSession, employee: EmployeeCreate):
     """
     Função que cria um novo funcionário
     """
     db_employee = EmployeeModel(**employee.model_dump())
     db.add(db_employee)
-    db.commit()
-    db.refresh(db_employee)
+    await db.commit()
+    await db.refresh(db_employee)
     return db_employee
 
-
-def delete_employee(db: Session, employee_id: int):
+async def delete_employee(db: AsyncSession, employee_id: int):
     """
     Função que deleta um funcionário
     """
-    db_employee = db.query(EmployeeModel).filter(EmployeeModel.employee_id == employee_id).first()
-    db.delete(db_employee)
-    db.commit()
+    db_employee = await get_employee(db, employee_id=employee_id)
+    if db_employee:
+        await db.delete(db_employee)
+        await db.commit()
     return db_employee
 
-
-def update_employee(db: Session, employee_id: int, employee: EmployeeUpdate):
+async def update_employee(db: AsyncSession, employee_id: int, employee: EmployeeUpdate):
     """
     Função que atualiza um funcionário
     """
-    db_employee = db.query(EmployeeModel).filter(EmployeeModel.employee_id == employee_id).first()
-
+    db_employee = await get_employee(db, employee_id=employee_id)
     if db_employee is None:
         return None
 
@@ -51,6 +49,6 @@ def update_employee(db: Session, employee_id: int, employee: EmployeeUpdate):
     for key, value in update_data.items():
         setattr(db_employee, key, value)
 
-    db.commit()
-    db.refresh(db_employee)
+    await db.commit()
+    await db.refresh(db_employee)
     return db_employee
